+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Schemas will be generated into own namespace `Schema`
 - RequestBodies and Responses will also be generated into their own namespace
+- date and date-time are supported, by default it is disabled and will be transformed into `string`.
 
 ## 0.2.0